+++ unlock-js/src/v7/index.js
@@ -8,6 +8,9 @@ import withdrawFromLock from './withdrawFromLock'
 import initializeTemplate from './initializeTemplate'
 import configureUnlock from './configureUnlock'
 import isLockManager from './isLockManager'
+import isKeyGranter from './isKeyGranter'
+import addKeyGranter from './addKeyGranter'
+import expireAndRefundFor from './expireAndRefundFor'
 
 export default {
   version: 'v7',
@@ -22,4 +25,7 @@ export default {
   withdrawFromLock,
   configureUnlock,
   isLockManager,
+  isKeyGranter,
+  addKeyGranter,
+  expireAndRefundFor,
 }